+++ NEWS.md
@@ -9,7 +9,7 @@
 * Small fix to parameter in `train()` #102 @jeffreyhanson
 * Small helper function for combining 2 different formula objects `combine_formulas()`
 * Small bug fixes dealing with `scenario()` projections and limits, plus unit tests #104
-* Fixed bug with adding `predictor_derivate()` to scenario predictors and added unit tests #106
+* Bug fixes with adding `predictor_derivate()` to scenario predictors and added unit tests #106
 
 # ibis.iSDM 0.1.2